backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
mysqlclient>=2.2.0
PyMySQL>=1.1.0
pytz>=2023.3
python-decouple>=3.8
gunicorn>=21.2.0
whitenoise>=6.6.0
